--- pkg/tsdb/cloudwatch/get_metric_data.go
@@ -1,191 +1,0 @@
-package cloudwatch
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"strconv"
-	"time"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/cloudwatch"
-	"github.com/grafana/grafana/pkg/components/null"
-	"github.com/grafana/grafana/pkg/components/simplejson"
-	"github.com/grafana/grafana/pkg/infra/metrics"
-	"github.com/grafana/grafana/pkg/tsdb"
-)
-
-func (e *CloudWatchExecutor) executeQuery(ctx context.Context, region string, queries map[string]*CloudWatchQuery, queryContext *tsdb.TsdbQuery) ([]*tsdb.QueryResult, error) {
-	queryResponses := make([]*tsdb.QueryResult, 0)
-
-	client, err := e.getClient(region)
-	if err != nil {
-		return queryResponses, err
-	}
-
-	params, err := buildGetMetricDataQuery(queries, queryContext)
-	if err != nil {
-		return queryResponses, err
-	}
-
-	nextToken := ""
-	mdr := make(map[string]map[string]*cloudwatch.MetricDataResult)
-	for {
-		if nextToken != "" {
-			params.NextToken = aws.String(nextToken)
-		}
-		resp, err := client.GetMetricDataWithContext(ctx, params)
-		if err != nil {
-			return queryResponses, err
-		}
-		metrics.MAwsCloudWatchGetMetricData.Add(float64(len(params.MetricDataQueries)))
-
-		for _, r := range resp.MetricDataResults {
-			if _, ok := mdr[*r.Id]; !ok {
-				mdr[*r.Id] = make(map[string]*cloudwatch.MetricDataResult)
-				mdr[*r.Id][*r.Label] = r
-			} else if _, ok := mdr[*r.Id][*r.Label]; !ok {
-				mdr[*r.Id][*r.Label] = r
-			} else {
-				mdr[*r.Id][*r.Label].Timestamps = append(mdr[*r.Id][*r.Label].Timestamps, r.Timestamps...)
-				mdr[*r.Id][*r.Label].Values = append(mdr[*r.Id][*r.Label].Values, r.Values...)
-			}
-		}
-
-		if resp.NextToken == nil || *resp.NextToken == "" {
-			break
-		}
-		nextToken = *resp.NextToken
-	}
-
-	for _, query := range queries {
-		queryRes := tsdb.NewQueryResult()
-		queryRes.RefId = query.RefId
-		queryRes.Meta = simplejson.New()
-		for i, stat := range query.Statistics {
-			lr := mdr[getQueryID(query, i)]
-			series, err := parseGetMetricDataTimeSeries(lr, query, *stat)
-			if err != nil {
-				return queryResponses, err
-			}
-			queryRes.Series = append(queryRes.Series, *series...)
-			queryRes.Meta.Set("searchExpressions", query.SearchExpressions)
-
-		}
-		queryResponses = append(queryResponses, queryRes)
-	}
-
-	return queryResponses, nil
-}
-
-func buildGetMetricDataQuery(queries map[string]*CloudWatchQuery, queryContext *tsdb.TsdbQuery) (*cloudwatch.GetMetricDataInput, error) {
-	startTime, err := queryContext.TimeRange.ParseFrom()
-	if err != nil {
-		return nil, err
-	}
-
-	endTime, err := queryContext.TimeRange.ParseTo()
-	if err != nil {
-		return nil, err
-	}
-
-	params := &cloudwatch.GetMetricDataInput{
-		StartTime: aws.Time(startTime),
-		EndTime:   aws.Time(endTime),
-		ScanBy:    aws.String("TimestampAscending"),
-	}
-
-	if !startTime.Before(endTime) {
-		return nil, fmt.Errorf("Invalid time range: Start time must be before end time")
-	}
-
-	for _, query := range queries {
-		query.SearchExpressions = []string{}
-		for i, stat := range query.Statistics {
-
-			// 1 minutes resolution metrics is stored for 15 days, 15 * 24 * 60 = 21600
-			if query.HighResolution && (((endTime.Unix() - startTime.Unix()) / int64(query.Period)) > 21600) {
-				return nil, errors.New("too long query period")
-			}
-
-			mdq := &cloudwatch.MetricDataQuery{
-				Id:         aws.String(getQueryID(query, i)),
-				ReturnData: aws.Bool(query.ReturnData),
-			}
-			if query.Expression != "" {
-				mdq.Expression = aws.String(query.Expression)
-			} else {
-				counter := 1
-				dimensionKeys := ""
-				searchTerm := fmt.Sprintf("MetricName=\"%v\" ", query.MetricName)
-				for key, values := range query.Dimensions {
-					dimensionKeys += fmt.Sprintf(",%s", key)
-					if len(values) == 1 && values[0] != "*" {
-						searchTerm += fmt.Sprintf("%s=(", key)
-						for i, value := range values {
-							searchTerm += fmt.Sprintf("\"%s\"", value)
-							if len(values) > 1 && i+1 != len(values) {
-								searchTerm += " OR "
-							}
-						}
-						searchTerm += ")"
-						if len(query.Dimensions) > 1 && counter != len(query.Dimensions) {
-							searchTerm += " AND "
-						}
-					}
-					counter++
-				}
-				searchExpression := fmt.Sprintf("SEARCH('{%s%s} %s', '%s', %s)", query.Namespace, dimensionKeys, searchTerm, *stat, strconv.Itoa(query.Period))
-				query.SearchExpressions = append(query.SearchExpressions, searchExpression)
-				mdq.Expression = aws.String(searchExpression)
-			}
-			params.MetricDataQueries = append(params.MetricDataQueries, mdq)
-
-		}
-	}
-	return params, nil
-}
-
-func parseGetMetricDataTimeSeries(lr map[string]*cloudwatch.MetricDataResult, query *CloudWatchQuery, stat string) (*tsdb.TimeSeriesSlice, error) {
-	result := tsdb.TimeSeriesSlice{}
-	for label, r := range lr {
-		if *r.StatusCode != "Complete" {
-			return &result, fmt.Errorf("Part of query is failed: %s", *r.StatusCode)
-		}
-
-		series := tsdb.TimeSeries{
-			Tags:   map[string]string{},
-			Points: make([]tsdb.TimePoint, 0),
-		}
-
-		for key, values := range query.Dimensions {
-			for _, value := range values {
-				if value == label {
-					series.Tags[key] = label
-				}
-			}
-		}
-
-		series.Name = formatAlias(query, stat, series.Tags, label)
-
-		for j, t := range r.Timestamps {
-			if j > 0 {
-				expectedTimestamp := r.Timestamps[j-1].Add(time.Duration(query.Period) * time.Second)
-				if expectedTimestamp.Before(*t) {
-					series.Points = append(series.Points, tsdb.NewTimePoint(null.FloatFromPtr(nil), float64(expectedTimestamp.Unix()*1000)))
-				}
-			}
-			series.Points = append(series.Points, tsdb.NewTimePoint(null.FloatFrom(*r.Values[j]), float64((*t).Unix())*1000))
-		}
-		result = append(result, &series)
-	}
-	return &result, nil
-}
-
-func getQueryID(query *CloudWatchQuery, statIndex int) string {
-	queryID := query.Identifier
-	if len(query.Statistics) > 1 {
-		queryID = query.Identifier + "_____" + strconv.Itoa(statIndex)
-	}
-	return queryID
-}
